requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.1.3
executing==2.0.1
frozenlist==1.4.0
h11==0.14.0
icecream==2.1.3
idna==3.4
importlib-metadata==6.8.0
libsql-client==0.3.0
lxml==4.9.3
multidict==6.0.4
outcome==1.3.0.post0
pyee==8.2.2
Pygments==2.16.1
pyppeteer==1.0.2
PySocks==1.7.1
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==1.26.18
websockets==10.4
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
